Changed the Stock and Base amounts of an Item to double instead of int. updated the DB to handle the change, fixed AddForm and editForm to handle these and fixed the display to only show 2 points after the decimal.

+++ lib/assets/item.dart
@@ -8,8 +8,8 @@ class Item {
   int id;
   String name;
   double pricePerUnit;
-  int amountInStock;
-  int amountBase;
+  double amountInStock;
+  double amountBase;
 
   Item(
       {this.id,
